ring broadcast packet format : 2 flits
git-svn-id: https://www.soclib.fr/svn/trunk/soclib@1146 2cf4c1f0-a32b-0410-9068-d80f8be93c97

+++ soclib/module/network_component/vci_simple_ring_network/caba/source/src/vci_simple_ring_network.cpp
@@ -64,7 +64,7 @@ tmpl(/**/)::VciSimpleRingNetwork( sc_module_name insname,
 		bool alloc_init = (x==0);             
 		std::ostringstream o;
 		o << name() << "_i_wrapper_" << x;
-		t_initiator_wrapper[x] = new soclib::caba::VciRingInitiatorWrapper<vci_param>(o.str().c_str(), alloc_init, wrapper_fifo_depth, mt,ringid, x);
+		t_initiator_wrapper[x] = new soclib::caba::VciRingInitiatorWrapper<vci_param>(o.str().c_str(), alloc_init, wrapper_fifo_depth, mt,ringid, (uint32_t) x);
 	}
     
         // generation wrappers targets
